Some cleaning + moving code to separate files
